refactor SetupGuide
